Fix gradle configrutaion
