+++ .vscode/settings.json
@@ -1,0 +1,16 @@
+{
+  "java.configuration.updateBuildConfiguration": "automatic",
+  "files.exclude": {
+    "**/.git": true,
+    "**/.svn": true,
+    "**/.hg": true,
+    "**/CVS": true,
+    "**/.DS_Store": true,
+    "bin/": true,
+    ".classpath": true,
+    ".project": true,
+    "java.home":true,
+    "java_home":"C:\\Program Files\\Java\\jdk-11.0.1"
+  }
+}
+//C:\Users\user\Documents\robot-2019>setx -m JAVA_HOME "C:\\Program Files\\Java\\jdk-11.0.1"
