+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,5 @@
-#Sun Jan 20 22:06:43 IST 2019
 distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
+distributionPath=permwrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-5.1.1-bin.zip
 zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-4.9-all.zip
+zipStorePath=permwrapper/dists
