+++ src/main/java/frc/robot/subsystems/drivetrain/pure_pursuit/Constants.java
@@ -1,4 +1,4 @@
-package robot.subsystems.drivetrain.pure_pursuit;
+package frc.robot.subsystems.drivetrain.pure_pursuit;
 
 public class Constants {
     //DRIVING CONSTANTS
