+++ src/main/java/frc/robot/subsystems/drivetrain/pure_pursuit/MainTest.java
@@ -1,4 +1,4 @@
-package robot.subsystems.drivetrain.pure_pursuit;
+package frc.robot.subsystems.drivetrain.pure_pursuit;
 
 public class MainTest {
 
@@ -6,9 +6,7 @@ public class MainTest {
     public static void main(String[] args) {
         Path path = new Path();
         path.appendWaypoint(new Waypoint(0, 0));
-        path.appendWaypoint(new Waypoint(0.9, 0));
-        path.appendWaypoint(new Waypoint(1.3, -0.5));
-        path.appendWaypoint(new Waypoint(5, 3));
+        path.appendWaypoint(new Waypoint(0.5, 0));
         System.out.println("BASIC PATH:");
         System.out.println(path);
         path.generateAll(Constants.WEIGHT_DATA, Constants.WEIGHT_SMOOTH, Constants.TOLERANCE, Constants.MAX_ACCEL, Constants.MAX_PATH_VELOCITY);
