+++ src/main/java/frc/robot/subsystems/drivetrain/pure_pursuit/Path.java
@@ -1,4 +1,4 @@
-package robot.subsystems.drivetrain.pure_pursuit;
+package frc.robot.subsystems.drivetrain.pure_pursuit;
 
 
 import java.util.ArrayList;
