+++ src/main/java/frc/robot/subsystems/drivetrain/pure_pursuit/Point.java
@@ -1,4 +1,4 @@
-package robot.subsystems.drivetrain.pure_pursuit;
+package frc.robot.subsystems.drivetrain.pure_pursuit;
 
 /**
  * A generic point class for storing (x, y) coordinates
