+++ src/main/java/frc/robot/subsystems/drivetrain/pure_pursuit/Vector.java
@@ -1,4 +1,4 @@
-package robot.subsystems.drivetrain.pure_pursuit;
+package frc.robot.subsystems.drivetrain.pure_pursuit;
 
 import edu.wpi.first.wpilibj.drive.Vector2d;
