--- src/main/java/robot/subsystems/drivetrain/Ports.java
@@ -1,13 +1,0 @@
-package robot.subsystems.drivetrain;
-
-public class Ports {
-    final static int leftForwardMotor = 8;
-    final static int leftBackMotor = 9;
-    final static int rightForwardMotor = 7;
-    final static int rightBackMotor = 6;
-    final static int rightEncoderChannelA = 2;
-    final static int rightEncoderChannelB = 3;
-    final static int leftEncoderChannelA = 0;
-    final static int leftEncoderChannelB = 1;
-
-}
